// example-project/tco-snt-virtualwarehouse-app/src/app/model/enums/CustomerCategory.ts
export enum CustomerCategory {
    //Участник СРП или сделки, заключенной в рамках СРП
    SHARING_AGREEMENT_PARTICIPANT = "SHARING_AGREEMENT_PARTICIPANT",
    EXPORTER = "EXPORTER",
    //Международный перевозчик
    TRANSPORTER = "TRANSPORTER",
    PUBLIC_OFFICE = "PUBLIC_OFFICE",
    NONRESIDENT = 'NONRESIDENT',
    PRINCIPAL = 'PRINCIPAL',
    JOINT_ACTIVITY_PARTICIPANT = 'JOINT_ACTIVITY_PARTICIPANT',
    LESSOR = 'LESSOR',
    FORWARDER = 'FORWARDER',
    BROKER = 'BROKER',
    COMMITTENT = 'COMMITTENT',
    LESSEE = 'LESSEE',
    RETAIL = 'RETAIL',
    INDIVIDUAL = 'INDIVIDUAL',
    LAWYER = 'LAWYER',
    BAILIFF = 'BAILIFF',
    MEDIATOR = 'MEDIATOR',
    NOTARY = 'NOTARY'
}